+++ contracts/cw721-base-ibc/src/query.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 
 use cosmwasm_std::{to_binary, Addr, Binary, BlockInfo, Deps, Env, Order, StdError, StdResult};
 
-use cw721::{
+use cw721_ibc::{
     AllNftInfoResponse, ApprovalResponse, ApprovalsResponse, ContractInfoResponse, CustomMsg,
     Cw721Query, Expiration, NftInfoResponse, NumTokensResponse, OperatorsResponse, OwnerOfResponse,
     TokensResponse,
@@ -93,7 +93,7 @@ where
 
         // token owner has absolute approval
         if token.owner == spender {
-            let approval = cw721::Approval {
+            let approval = cw721_ibc::Approval {
                 spender: token.owner.to_string(),
                 expires: Expiration::Never {},
             };
@@ -105,7 +105,7 @@ where
             .into_iter()
             .filter(|t| t.spender == spender)
             .filter(|t| include_expired || !t.is_expired(&env.block))
-            .map(|a| cw721::Approval {
+            .map(|a| cw721_ibc::Approval {
                 spender: a.spender.into_string(),
                 expires: a.expires,
             })
@@ -133,7 +133,7 @@ where
             .approvals
             .into_iter()
             .filter(|t| include_expired || !t.is_expired(&env.block))
-            .map(|a| cw721::Approval {
+            .map(|a| cw721_ibc::Approval {
                 spender: a.spender.into_string(),
                 expires: a.expires,
             })
@@ -281,8 +281,8 @@ where
     }
 }
 
-fn parse_approval(item: StdResult<(Addr, Expiration)>) -> StdResult<cw721::Approval> {
-    item.map(|(spender, expires)| cw721::Approval {
+fn parse_approval(item: StdResult<(Addr, Expiration)>) -> StdResult<cw721_ibc::Approval> {
+    item.map(|(spender, expires)| cw721_ibc::Approval {
         spender: spender.to_string(),
         expires,
     })
@@ -292,7 +292,7 @@ fn humanize_approvals<T>(
     block: &BlockInfo,
     info: &TokenInfo<T>,
     include_expired: bool,
-) -> Vec<cw721::Approval> {
+) -> Vec<cw721_ibc::Approval> {
     info.approvals
         .iter()
         .filter(|apr| include_expired || !apr.is_expired(block))
@@ -300,8 +300,8 @@ fn humanize_approvals<T>(
         .collect()
 }
 
-fn humanize_approval(approval: &Approval) -> cw721::Approval {
-    cw721::Approval {
+fn humanize_approval(approval: &Approval) -> cw721_ibc::Approval {
+    cw721_ibc::Approval {
         spender: approval.spender.to_string(),
         expires: approval.expires,
     }
