+++ packages/cw721-ibc/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "cw721"
+name = "cw721-ibc"
 version = "0.13.2"
 authors = [
     "Ethan Frey <ethanfrey@users.noreply.github.com>",
     "Orkun Külçe <orkun@deuslabs.fi>",
+    "Michael Scotto <m@publicawesome.com>"
 ]
 edition = "2018"
 description = "Definition and types for the CosmWasm-721 NFT interface"
