+++ packages/cw721-ibc/examples/schema.rs
@@ -2,12 +2,15 @@ use std::env::current_dir;
 use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, export_schema_with_title, remove_schemas, schema_for};
+use cosmwasm_std::Empty;
 
-use cw721::{
-    AllNftInfoResponse, ApprovalResponse, ApprovalsResponse, ContractInfoResponse, NftInfoResponse,
-    NumTokensResponse, OperatorsResponse, OwnerOfResponse, TokensResponse,
+use cw721_ibc::{
+    AllNftInfoResponse, ApprovalResponse, ApprovalsResponse, ContractInfoResponse, Cw721ExecuteMsg,
+    Cw721QueryMsg, Cw721ReceiveMsg, NftInfoResponse, NumTokensResponse, OperatorsResponse,
+    OwnerOfResponse, TokensResponse,
 };
-use cw721_ibc::{ExecuteMsg, Extension, InstantiateMsg, MinterResponse, QueryMsg};
+
+type Extension = Option<Empty>;
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
@@ -15,9 +18,9 @@ fn main() {
     create_dir_all(&out_dir).unwrap();
     remove_schemas(&out_dir).unwrap();
 
-    export_schema(&schema_for!(InstantiateMsg), &out_dir);
-    export_schema_with_title(&schema_for!(ExecuteMsg<Extension>), &out_dir, "ExecuteMsg");
-    export_schema(&schema_for!(QueryMsg), &out_dir);
+    export_schema(&schema_for!(Cw721ExecuteMsg), &out_dir);
+    export_schema(&schema_for!(Cw721QueryMsg), &out_dir);
+    export_schema(&schema_for!(Cw721ReceiveMsg), &out_dir);
     export_schema_with_title(
         &schema_for!(AllNftInfoResponse<Extension>),
         &out_dir,
@@ -27,13 +30,12 @@ fn main() {
     export_schema(&schema_for!(ApprovalsResponse), &out_dir);
     export_schema(&schema_for!(OperatorsResponse), &out_dir);
     export_schema(&schema_for!(ContractInfoResponse), &out_dir);
-    export_schema(&schema_for!(MinterResponse), &out_dir);
+    export_schema(&schema_for!(OwnerOfResponse), &out_dir);
     export_schema_with_title(
         &schema_for!(NftInfoResponse<Extension>),
         &out_dir,
         "NftInfoResponse",
     );
     export_schema(&schema_for!(NumTokensResponse), &out_dir);
-    export_schema(&schema_for!(OwnerOfResponse), &out_dir);
     export_schema(&schema_for!(TokensResponse), &out_dir);
 }
